+++ tests/test_speed_api_dnld.py
@@ -4,19 +4,17 @@
 from sys import getsizeof
 from timeit import default_timer
 
-from pmidcite.icite.pmid_dnlder import NIHiCiteDownloader
-
-from tests.icite import DIR_ICITE
 from tests.icite import dir_icite_clobber
 from tests.icite import dir_icite_wc_l
 from tests.prt_hms import prt_hms
 from tests.pmids_i3 import PMIDS
+from tests.utils import init_dnldr
 
 
 def test_speed_api_dnld():
     """Test speed for download NIH citation data"""
     force_dnld = True
-    dnldr = _init_dnldr(force_dnld)
+    dnldr = init_dnldr(force_dnld)
 
     num = 5000
     print(f'tests.pmids_i3 len(PMIDS)={len(PMIDS)}')
@@ -24,7 +22,6 @@ def test_speed_api_dnld():
     pmids = PMIDS
     # _dnld_icite_v_icites(dnldr, pmids)
 
-    dir_icite_clobber()
     ## nihentries = dnldr.api.dnld_icites(pmids)
     ## nihentries = dnldr.get_pmid2paper(pmids)
     ## nihentries = dnldr.get_icites(pmids)
@@ -45,15 +42,6 @@ def _dnld_icite_v_icites(dnldr, pmids):
             print(f"Downloading item {idx:,}")
         dnldr.api.dnld_icite(pmid)  # NIHiCiteEntry
     tic = prt_hms(tic, f"Downloaded {num_pmids} items w/dnld_icite")
-
-
-def _init_dnldr(force_dnld):
-    """Initialize a NIH Downloader and tmp dir, src/tests/icite"""
-    tic = default_timer()
-    dir_icite_clobber(prt=None)
-    dnldr = NIHiCiteDownloader(DIR_ICITE, force_dnld)
-    tic = prt_hms(tic, "Initialize NIH citation downloader")
-    return dnldr
 
 
 if __name__ == '__main__':